refactor activation to screens and handle lifecycle
